Split shell client liveness tests
